refactor(web): StructuredOutputSchema Code Splitting
