+++ web/src/views/Workflow/components/Properties/ModelConfig/StructuredOutputSchemaModal.tsx
@@ -1,11 +1,10 @@
 /*
- * @Author: ZhaoYing 
- * @Date: 2026-06-09 
+ * @Author: ZhaoYing
+ * @Date: 2026-06-09
  * Structured Output Schema Modal
  */
-
 import { forwardRef, useImperativeHandle, useState, useMemo, useRef } from 'react';
-import { Button, Input, Select, Tooltip, Flex, Space, Divider, Tree, Form, App, Switch } from 'antd';
+import { Button, Flex, Space, Tree, Form, App } from 'antd';
 import { DownOutlined } from '@ant-design/icons';
 import type { DataNode } from 'antd/es/tree';
 import { useTranslation } from 'react-i18next';
@@ -14,53 +13,13 @@ import RbModal from '@/components/RbModal';
 import PageTabs from '@/components/PageTabs';
 import CodeMirrorEditor from '@/components/CodeMirrorEditor';
 import JsonImportModal, { type JsonImportModalRef } from './JsonImportModal';
-import clsx from 'clsx';
-
-export interface StructuredOutputSchemaModalRef {
-  handleOpen: (schema: JsonSchema) => void;
-  handleClose: () => void;
-}
-
-export interface Field {
-  name?: string;
-  type: string;
-  children?: Field[];
-  description?: string;
-  required?: boolean;
-}
-
-export type JsonSchema = Field[];
-
-const defaultJsonSchema: JsonSchema = []
-
-const typeOptions = [
-  { value: 'string', label: 'string' },
-  { value: 'number', label: 'number' },
-  { value: 'boolean', label: 'boolean' },
-  { value: 'object', label: 'object' },
-  { value: 'array[string]', label: 'array[string]' },
-  { value: 'array[number]', label: 'array[number]' },
-  { value: 'array[object]', label: 'array[object]' },
-];
+import FieldRow from './FieldRow';
+import type { JsonSchema, StructuredOutputSchemaModalRef, Field } from './types';
+import { defaultJsonSchema, getAllObjectKeys } from './constant'
 
 interface StructuredOutputSchemaModalProps {
   refresh: (schema: JsonSchema) => void;
 }
-
-// 默认展开所有 object 类型节点的 key
-const getAllObjectKeys = (list: Field[], path: number[] = []): string[] => {
-  const keys: string[] = [];
-  list.forEach((field, index) => {
-    const currentPath = [...path, index];
-    if (field.type.includes('object')) {
-      keys.push(currentPath.join(','));
-      if (field.children) {
-        keys.push(...getAllObjectKeys(field.children, currentPath));
-      }
-    }
-  });
-  return keys;
-};
 
 export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModalRef, StructuredOutputSchemaModalProps>(({
   refresh,
@@ -105,7 +64,7 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
 
     let newFields: Field[];
     let newFieldIndexPath: number[];
-    
+
     // 如果 parentIndexPath 为空，直接添加到根级别
     if (parentIndexPath.length === 0) {
       newFields = [...fields, newField];
@@ -126,7 +85,7 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
           };
         });
       };
-      
+
       // 获取父级字段，确定新字段的位置
       const getFieldByIndexPath = (list: Field[], indices: number[]): Field | null => {
         let current: Field | null = null;
@@ -138,15 +97,15 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
         }
         return current;
       };
-      
+
       const parentField = getFieldByIndexPath(fields, parentIndexPath);
       const parentChildrenCount = parentField?.children?.length || 0;
-      
+
       newFields = updateByIndexPath(fields, parentIndexPath, (field) => ({
         ...field,
         children: [...(field.children || []), newField]
       }));
-      
+
       newFieldIndexPath = [...parentIndexPath, parentChildrenCount];
 
       // 展开父节点
@@ -164,13 +123,13 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
   /** Delete a field by index path (e.g., [0, 2] means the 3rd child of the 1st root field) */
   const deleteField = (indexPath: number[]) => {
     if (indexPath.length === 0) return;
-    
+
     const deleteFromList = (list: Field[], indices: number[]): Field[] => {
       if (indices.length === 1) {
         // Always allow deletion
         return list.filter((_, i) => i !== indices[0]);
       }
-      
+
       const [first, ...rest] = indices;
       return list.map((field, i) => {
         if (i !== first) return field;
@@ -182,20 +141,33 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
         };
       });
     };
-    
+
     const newFields = deleteFromList(fields, indexPath);
     setFields(newFields);
   };
 
   const toggleEdit = (id: string) => {
     console.log('toggleEdit', editingId, id )
-    
+
     setEditingId(editingId === id ? null : id);
     if (editingId !== id) {
       setTimeout(() => {
         editForm.resetFields();
       }, 0);
     }
+  };
+
+  /** Exit edit mode, removing the field if it was a freshly added (unnamed) one */
+  const handleCancelField = (indexPath: number[], field: Field) => {
+    if (!field.name) {
+      deleteField(indexPath);
+      setEditingId(null);
+    } else {
+      setEditingId(null);
+    }
+    setTimeout(() => {
+      editForm.resetFields();
+    }, 0);
   };
 
   /** Save a field by validating and committing its current form values, then exit edit mode */
@@ -223,17 +195,17 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
 
     // Find parent field and siblings using index path
     const parentIndexPath = indexPath.slice(0, -1);
-    const parentField = parentIndexPath.length > 0 
-      ? getFieldByIndexPath(fields, parentIndexPath) 
+    const parentField = parentIndexPath.length > 0
+      ? getFieldByIndexPath(fields, parentIndexPath)
       : null;
     const siblings = parentIndexPath.length === 0
       ? fields.map((f) => f.name)
       : (parentField?.children || []).map((c) => c.name);
-    
+
     // Get old name from the field being edited
     const editedField = getFieldByIndexPath(fields, indexPath);
     const oldName = editedField?.name;
-    
+
     // Validation: name must be unique within its parent (excluding the field itself)
     const conflict = siblings.some((n) => n === newName && n !== oldName);
     if (conflict) {
@@ -270,170 +242,26 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
     setEditingId(null);
   };
 
-  /** Render a single field row (used recursively for nested fields) */
-  const renderFieldRow = (field: Field, indexPath: number[] = []): React.ReactNode => {
-    const currentIndexPath = indexPath.join(',');
-    const isEditing = editingId === currentIndexPath;
-    const isObject = field.type.includes('object');
-
-    return (
-      <div
-        onClick={(e) => e.stopPropagation()}
-        className={clsx(
-          'rb:flex rb:items-center rb:gap-4 rb:py-1 rb:px-3 rb:rounded-lg',
-          {
-            'rb-border rb:bg-white': isEditing
-          }
-        )}
-      >
-        {isEditing ? (
-          <Form
-            form={editForm}
-            layout="inline"
-            className="rb:flex-1"
-            initialValues={{
-              name: !field.name ? undefined : field.name,
-              type: field.type,
-              description: field.description,
-              required: field.required
-            }}
-          >
-            <Flex justify="space-between" className="rb:w-full!">
-              <Space>
-                <Form.Item name="name" noStyle>
-                  <Input
-                    size="small"
-                    placeholder={t('workflow.config.llm.fieldName')}
-                    variant="borderless"
-                  />
-                </Form.Item>
-                <Form.Item name="type" noStyle>
-                  <Select
-                    size="small"
-                    options={typeOptions}
-                    className="rb:w-34!"
-                    variant="borderless"
-                  />
-                </Form.Item>
-              </Space>
-
-              <Flex align="center" gap={8}>
-                <Form.Item name="required" noStyle valuePropName="checked">
-                  <Switch
-                    size="small"
-                    checkedChildren={t('workflow.config.llm.required')}
-                    unCheckedChildren={t('workflow.config.llm.unRequired')}
-                    onClick={(_, e) => e.stopPropagation()}
-                  />
-                </Form.Item>
-
-                <Button
-                  size="small"
-                  className="rb:text-[12px]!"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    // If the field is a newly added field (name is undefined), remove it from
-                    // the schema on cancel to keep the data clean.
-                    if (!field.name) {
-                      deleteField(indexPath);
-                      setEditingId(null);
-                    } else {
-                      setEditingId(null);
-                    }
-
-                    setTimeout(() => {
-                      editForm.resetFields();
-                    }, 0);
-                  }}
-                >{t('common.cancel')}</Button>
-
-                <Button
-                  size="small"
-                  type="primary"
-                  className="rb:text-[12px]!"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    handleSaveField(indexPath);
-                  }}
-                >
-                  {t('common.confirm')}
-                </Button>
-              </Flex>
-            </Flex>
-
-            <Form.Item name="description" className="rb:flex-1 rb:mb-0!">
-              <Input
-                size="small"
-                variant="borderless"
-                placeholder={t('workflow.config.llm.addDescription')}
-              />
-            </Form.Item>
-          </Form>
-        ) : (
-          <Flex justify="space-between" gap={8} className="rb:w-full!">
-            <div>
-              <Flex align="center" gap={8} className="rb:shrink-0">
-                <span className="rb:text-sm rb:font-medium">{field.name || t('workflow.config.llm.fieldName')}</span>
-                <span className="rb:text-xs rb:text-[#5B6167]">{field.type}</span>
-                {field.required && (
-                  <span className="rb:text-xs rb:text-[#F5222D]">{t('workflow.config.llm.required')}</span>
-                )}
-              </Flex>
-
-              {field.description &&
-                <div className="rb:flex-1 rb:text-[#5B6167]">
-                  {field.description}
-                </div>
-              }
-            </div>
-
-            <div className="rb:flex rb:items-center rb:gap-2">
-              {isObject && (
-                <Tooltip title={t('workflow.config.llm.addField')}>
-                  <div
-                    className="rb:size-4.5 rb:cursor-pointer rb:bg-cover rb:bg-[url('@/assets/images/common/plus_light_grey.svg')]"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      addField(indexPath);
-                    }}
-                  />
-                </Tooltip>
-              )}
-
-              <Tooltip title={t('common.edit')}>
-                <div
-                  className="rb:size-4.5 rb:cursor-pointer rb:bg-cover rb:bg-[url('@/assets/images/common/edit.svg')]"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    e.preventDefault();
-                    toggleEdit(currentIndexPath);
-                  }}
-                ></div>
-              </Tooltip>
-
-              <Tooltip title={t('common.delete')}>
-                <div
-                  className="rb:size-4.5 rb:cursor-pointer rb:bg-cover rb:bg-[url('@/assets/images/common/delete_red.svg')]"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    deleteField(indexPath);
-                  }}
-                ></div>
-              </Tooltip>
-            </div>
-          </Flex>
-        )}
-      </div>
-    );
-  };
-
   /** Recursively convert fields to Tree data */
   const fieldsToTreeData = (fieldsList: Field[], parentPath: (string | undefined)[] = [], indexPath: number[] = []): DataNode[] => {
     return fieldsList.map((field, index) => {
       const currentIndexPath = [...indexPath, index];
       return {
         key: currentIndexPath.join(','),
-        title: renderFieldRow(field, currentIndexPath),
+        title: (
+          <FieldRow
+            field={field}
+            indexPath={currentIndexPath}
+            editingId={editingId}
+            isEditing={editingId === currentIndexPath.join(',')}
+            editForm={editForm}
+            onAddField={addField}
+            onDeleteField={deleteField}
+            onToggleEdit={toggleEdit}
+            onSaveField={handleSaveField}
+            onCancel={handleCancelField}
+          />
+        ),
         children: field.type.includes('object') && field.children
           ? fieldsToTreeData(field.children, [...parentPath, field.name], currentIndexPath)
           : undefined
@@ -450,10 +278,6 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
     setFields(next);
   };
 
-  const syncToJson = () => {
-    // TODO: Implement sync to JSON logic
-  };
-
   /** Derive fields and build tree data + field map in one pass */
   const treeData: DataNode[] = useMemo(() => {
     const walk = (list: Field[]) => {
@@ -462,6 +286,7 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
       });
     };
     walk(fields);
+    if (editingId) return fieldsToTreeData(fields);
     return [
       ...fieldsToTreeData(fields),
       {
@@ -476,7 +301,6 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
     handleClose
   }));
 
-  console.log('treeData', treeData)
   return (
     <>
     <RbModal
@@ -551,13 +375,6 @@ export const StructuredOutputSchemaModal = forwardRef<StructuredOutputSchemaModa
             language="json"
             variant="outlined"
             height="320px"
-            // onChange={(value) => {
-            //   try {
-            //     setFields(JSON.parse(value));
-            //   } catch (error) {
-            //     console.error('Invalid JSON', error);
-            //   }
-            // }}
             placeholder={t('workflow.config.llm.jsonSchemaPlaceholder')}
           />
         </div>
